Update main.rs
Added Delay argument as just any whole number passed. Default delay if not passed is flat 50ms
Added data val to arguments, all arguments are now converted to lowercase to avoid capital letters breaking macro
changed "no_ctrl" to "press" so the keypress is managed for you.
Changed the way the macro is run to shorten the code.
Changed the way arguments are handled so any argument can be in any order. E.G "resupply 10 press" will also work as "10 press resupply" etc.
Added in some more stratagems as well as fixing the EMS mortar

Added a bunch of comments to the code for future maintance